boto_tutorial/requirements.txt
fabric2
python-dotenv
